refactored fs
