+++ .vscode/arduino.json
@@ -1,0 +1,5 @@
+{
+    "board": "arduino:avr:uno",
+    "programmer": "avrisp",
+    "sketch": "software\\main.ino"
+}
