+++ README.md
@@ -1,1 +1,5 @@
 # ECE1895 - Bop It
+### Hardware folder
+- contains altium, spice, and other hardware related files
+### Software folder
+- contains c/c++ and arduino files for microcontroller
