+++ software/main.ino
@@ -1,0 +1,8 @@
+#include "slotmachinebopit.c"
+
+void setup() {
+
+}
+void loop() {
+
+}
